% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2ctgr.R
\encoding{UTF-8}
\name{add2ctgr}
\alias{add2ctgr}
\title{Enviar nuevas observaciones a intervalos ya existentes}
\usage{
add2ctgr(x, intrvls)
}
\arguments{
\item{x}{Un vector de números}

\item{intrvls}{Un objeto devuelto por la función \code{num2ctgr}}
}
\value{
Una lista con nombres que contiene los intervalos (\code{intrvls})
    y los cortes usados para crearlos (\code{brks})
}
\description{
Dado un objeto retornado por \code{num2ctgr}, al ingresar
    un vector numérico en esta función se asignarán a dichos números
    los intervalos correspondientes que se encuentren en \code{intrvls}.
    Si sucede que un número sea inferior (o mayor) al intervalo con
    los números más pequeños (o grandes) se asumirá pertenencia al
    primero (ultimó) de ellos.
}
\examples{
example_vect <- c(-1,-1,-10,1,0,2,3,10,5,11,12,15,20,-4)
example_int <- num2ctgr(x = example_vect, type = "jenks", n = 2)
example_vect2 <- c(1,10,100,-20)
add2ctgr(example_vect2, example_int)
}
