% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{add_date_split}
\alias{add_date_split}
\title{Adicionar columnas relacionadas con variables de tiempo}
\usage{
add_date_split(df, date_col, rm_date = TRUE, ...)
}
\arguments{
\item{df}{El \code{data.frame} a ser extendido.}

\item{date_col}{El nombre de la columna que contiene la información de fechas.}

\item{rm_date}{Eliminar o no la columna de fechas.}

\item{...}{Una serie de pares de clave/valor donde los últimos son
\code{TRUE} (por defecto) o \code{FALSE}. Las claves corresponden con
los campos que salida de la función, en este caso son:
\describe{
     \item{\code{year}:}{Devolver el año}
     \item{\code{month}:}{Devolver el mes númerico}
     \item{\code{week}:}{Devolver el número de la semana}
     \item{\code{week_day}:}{Devolver el día de la semana}
     \item{\code{year}:}{Devolver el año}
}}
}
\value{
Un \code{data.frame} aumentado con las columnas seleccionadas.
}
\description{
Utilidades de pre-procesamiento para modelos, cuya finalidad es
    crear nuevas columnas a partir de información condensada en una
    correspondiente a fechas.
}
\examples{
# Creacion de un data.frame con una columna de fechas
X <- data.frame(FECHA = seq(as.Date("2020-01-01"), length = 10,
         by = "1 day"), COLOR = rep(c("Azul", "Verde"), 5))
# Aumento de columnas
add_date_split(X,"FECHA")
}
