% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{add_laggs}
\alias{add_laggs}
\title{Adición de rezagos a la matriz de características/diseño.}
\usage{
add_laggs(X, Y, p = 10)
}
\arguments{
\item{X}{La matriz de características.}

\item{Y}{El vector de respuestas.}

\item{p}{El rezago máximo de Y que se adicionará a X.}
}
\value{
Una matriz X aumentada con valores rezagados de Y.
}
\description{
Extensión de \code{lagpad} que ayuda al diseño de
    modelos de regresión con términos autoregresivos en X. Dado que
    \code{lagpad} devuelve vectores con NA, el procedimiento las elimina
    a favor de la integridad del modelo final, no obstante, debe
    considerarse esto dentro de los análisis pertinentes.
}
\examples{
# Simular un AR(1)
Y <- stats::arima.sim(list(order(p=0.7)),n = 100)
# Convertir a numerico
Y <- as.numeric(Y)
# Crear una matriz X de dos columnas
X <- data.frame(color = sample(c("Azul","Verde"), 100, replace = TRUE))
# Adicion de lags
X <- add_laggs(X, Y, p = 1)
}
