% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{add_spatial_f}
\alias{add_spatial_f}
\title{Adicionar columnas relacionadas con agrupaciones espaciales}
\usage{
add_spatial_f(df, sf_obj, ...)
}
\arguments{
\item{df}{El \code{data.frame} a ser extendido.}

\item{sf_obj}{El objeto espacial usado como referencia.}

\item{...}{Los nombres de las columnas a preservar de \code{sf_obj}.}
}
\value{
Un \code{data.frame} aumentado con las columnas seleccionadas.
}
\description{
Utilidad de pre-procesamiento para modelos, cuya finalidad es
    aumentar el conjunto de datos con columnas de información espacial
    la cual es obtenida desde un objeto del tipo \code{sf} o \code{sp} cuyas
    geometrías corresponden a polígonos bien definidos. Se debe garantizar
    que el objeto a ser aumentado tenga en sus filas la información
    correspondiente a puntos espaciales (latitud y longitud) usando nombres
    apropiados para dichas columnas y que todo punto este contenido en
    alguna de las geometrías del objeto espacial.
}
\examples{
# Usando como ejemplo Bogota, puntos localizados en centros de localidades
# Puente Aranda, Teusaquillo y Kennedy
X <- data.frame(lng = c(-74.1066, -74.0888, -74.1533),
         lat = c(4.61253, 4.63708, 4.6435))
# Incluir el codigo y nombre de las localidades
add_spatial_f(X, bogota, "CODIGO_DIV1", "NOMBRE_DIV1")
}
