% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_mod_selector.R
\encoding{UTF-8}
\name{ar_mod_selector}
\alias{ar_mod_selector}
\title{Ajuste automático de modelos AR con validación por RMSE.}
\usage{
ar_mod_selector(
  split_out,
  p_max = 5,
  dates_col = NULL,
  factor_split = NULL,
  quiet = FALSE
)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.}

\item{p_max}{El máximo rezago a usar al explorar modelos.}

\item{dates_col}{El nombre que tienen la columna de fechas. Opcional.}

\item{factor_split}{El nombre de las columnas que crean las particiones
por factores. Ideal para múltiples series de tiempo, no necesariamente
en una misma ventana de tiempo.}

\item{quiet}{Callar las salidas en consola de los modelos.}
}
\value{
Una lista que incluye el mejor modelo y los RMSE en entrenamiento
    y validación (un pase)
}
\description{
Modelos AR automáticos para una serie de tiempo o múltiples
    que son identificadas por un conjunto de factores únicos (datos
    longitudinales).

    No se realizan múltiples pases de validación.

    TODO: Ejemplos y múltiples pases de validación.
}
