% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{arima_lin}
\alias{arima_lin}
\alias{arima_lin_metrics}
\alias{lin_arima_metrics}
\title{Medidas de rendimiento para el modelo experimental de dos etapas
    ARIMA+Modelo lineal}
\usage{
arima_lin_metrics(
  arima_fit,
  lin_fit,
  split_y,
  dates_col,
  objective = "validation",
  ...
)

lin_arima_metrics(
  arima_fit,
  lin_fit,
  split_y,
  dates_col,
  objective = "validation",
  ...
)
}
\arguments{
\item{arima_fit}{Una lista de modelos ARIMA (residuales u observaciones)}

\item{lin_fit}{El modelo de regresión (residuales u observaciones)}

\item{split_y}{La partición del conjunto de datos en
entrenamiento/validación/prueba}

\item{dates_col}{El nombre de la columna que contiene la fechas}

\item{objective}{Seleccionar entrenamiento "train" o validación "validation"}

\item{...}{Los nombres de las medidas adicionales, entre las que se
encuentran:
\describe{
    \item{\code{all}}{todas las metricas mencionadas enseguida.}
    \item{\code{mean}}{media de los residuos.}
    \item{\code{mean_observed}}{media de la variable respuesta.}
    \item{\code{mae}}{error absoluto medio de los residuos.}
    \item{\code{cve_rmse}}{cociente entre \code{rmse} y \code{mean}.}
    \item{\code{sd_observed}}{desviación estándar de la variable respuesta.}
    \item{\code{rmse_scaled}}{RMSE obtenido con residuales estandarizados por la suma absoluta de las diferencias de las observaciones.}
}}
}
\value{
Una lista de medidas de rendimiento dentro de muestra o fuera,
    según la elección del usuario
}
\description{
Usada para obtener diferentes medidas para evaluación de
    rendimiento sobre modelos definidos por el usuario. La medición
    se realiza sobre el conjunto de entrenamiento o validación.
    Se asume que X existe y es proporcionada. Además, se requiere el
    ingreso de un conjunto de datos con información de residuales
    (no necesario para el conjunto de pruebas)

    Por defecto entrega el RMSE.

    TODO: Ejemplos.
}
