% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_lin_selector.R
\encoding{UTF-8}
\name{arima_lin_selector}
\alias{arima_lin_selector}
\title{Ajuste de modelo experimental en dos etapas ARIMA + lineal
    con validación por RMSE.}
\usage{
arima_lin_selector(
  split_out,
  factor_split = NULL,
  dates_col = NULL,
  step_selection = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.}

\item{factor_split}{Los nombres de las columnas usadas para identificar
los objetos medidos en el tiempo.}

\item{dates_col}{El nombre de la columna que contiene las fechas.}

\item{step_selection}{¿Se debe probar subconjuntos de variables para el
modelo lineal de los residuales?}

\item{quiet}{Callar las salidas en consola de los modelos.}
}
\value{
Una lista que incluye los modelos ARIMA elegidos (\code{arima_fit}),
    el modelo lineal de los residuales (\code{lin_fit}), y los
    RMSE conjuntos obtenidos en entrenamiento y validación
}
\description{
Este es un algoritmo experimental que corre en primer lugar
    los modelos ARIMA automáticos (\code{arima_mod_selector})
    para el conjunto de datos solicitado
    y a los cuales se les extrae sus residuales para ajustar un modelo
    de regresión lineal (\code{lin_mod_selector}). La variable obtenida
    es una combinación lineal del pronosticó ARIMA más el residual
    estimado.

    No se realizan múltiples pases de validación.

    TODO: Ejemplos y múltiples pases de validación.
}
