% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_mod_selector.R
\encoding{UTF-8}
\name{arima_mod_selector}
\alias{arima_mod_selector}
\title{Ajuste de modelos ARIMA con validación por RMSE.}
\usage{
arima_mod_selector(
  split_out,
  dates_col = NULL,
  factor_split = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.}

\item{dates_col}{El nombre de la columna que contiene las fechas.}

\item{factor_split}{Los nombres de las columnas usadas para identificar
los objetos medidos en el tiempo.}

\item{quiet}{Callar las salidas en consola de los modelos.}

\item{...}{Argumentos adicionales para componete estacional}
}
\value{
Una lista que incluye el mejor modelo ARIMA (para una serie) ó
    una lista que incluye los factores que identifican cada serie
    \code{factor_list} y una lista de los mejores modelos para cada
    serie individual. Las salidas de modelos incluyen los RMSE en
    entrenamiento y validación.
}
\description{
Modelos ARIMA automáticos para una serie de tiempo o múltiples
    que son identificadas por un conjunto de factores únicos (datos
    longitudinales).

    No se realizan múltiples pases de validación.

    TODO: Ejemplos y múltiples pases de validación.
}
