% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{arimax_metrics}
\alias{arimax_metrics}
\title{Medidas de rendimiento para modelos ARIMAX}
\usage{
arimax_metrics(
  model_list,
  split_out,
  xreg_cols,
  fourier_K = NULL,
  dates_col = NULL,
  factor_split = NULL,
  objective = "validation"
)
}
\arguments{
\item{model_list}{Lista de modelos o modelo único al que se requiere
extrar las medidas de rendimiento un paso adelante}

\item{split_out}{La partición del conjunto de datos en
entrenamiento/validación/prueba}

\item{xreg_cols}{El nombre de las columnas usadas como regresoras exóngenas}

\item{fourier_K}{El grado del suavizado de fourier usado. NULL para omitir}

\item{dates_col}{El nombre que lleva la columna de fechas}

\item{factor_split}{El nombre de las columnas usado para realizar las
particiones y asignación de los modelos}

\item{objective}{Seleccionar entrenamiento "train" o validación "validation"}
}
\value{
Salidas de rendimiento para los modelos ARIMAX suministrados
}
\description{
Usada para obtener diferentes medidas para evaluación de
    rendimiento sobre modelos definidos por el usuario. La medición
    se realiza sobre el conjunto de validación.

    Por defecto entrega el RMSE.

    TODO: Ejemplos.
}
