% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{available_dates}
\alias{available_dates}
\title{Fechas disponibles para Londres.}
\usage{
available_dates(static = FALSE)
}
\arguments{
\item{static}{Bolean, falso por defecto. En caso de ser
verdadero se devuelven las fechas disponibles del
conjunto de datos estático TODO.}
}
\value{
Listado de fechas formateadas año y mes.
}
\description{
Fechas disponibles para Londres.
}
