% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_mod_selector.R
\encoding{UTF-8}
\name{Bcoef}
\alias{Bcoef}
\title{Extracción de coeficientes de objeto VAR}
\usage{
Bcoef(x)
}
\arguments{
\item{x}{Un objeto de clase \code{varest}}
}
\value{
Matriz de coeficientes VAR
}
\description{
Objeto exportado del paquete \code{vars}
}
