% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_col_crim.R
\docType{data}
\encoding{UTF-8}
\name{bog_hurto_celulares}
\alias{bog_hurto_celulares}
\title{Reportes de robos de celulares en Bogotá}
\format{
An object of class \code{data.frame} with 215673 rows and 7 columns.
}
\source{
SIEDCO
}
\usage{
bog_hurto_celulares
}
\description{
Reportes de robos de celulares en Bogotá
}
\keyword{datasets}
