% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_col_crim.R
\docType{data}
\encoding{UTF-8}
\name{bog_hurto_motos}
\alias{bog_hurto_motos}
\title{Reportes simulados de robos de motos en Bogotá}
\format{
An object of class \code{data.frame} with 60566 rows and 7 columns.
}
\source{
Elaboración propia
}
\usage{
bog_hurto_motos
}
\description{
Reportes simulados de robos de motos en Bogotá
}
\keyword{datasets}
