% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{bogota}
\alias{bogota}
\title{Informaci칩n espacial de Bogot치}
\format{
TODO: Revisar el archivo.
}
\source{
[Datos abiertos Bogot치](https://datosabiertos.bogota.gov.co/dataset/unidad-de-planeamiento-zonal-bogota-d-c)
}
\usage{
bogota
}
\description{
Un \code{SpatialPolygonsDataFrame} del paquete \code{sp}
    que contiene poligonos usados para la ciudad de Bogot치 sub-dividida
    en UPZ y Localidades.
}
\keyword{datasets}
