% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{cartagena}
\alias{cartagena}
\title{Información espacial de Cartagena}
\format{
TODO: Revisar el archivo.
}
\source{
[Invemar](http://geonodesiam.invemar.org.co/layers/geonode%3Abarrios_pot_ctgna)
}
\usage{
cartagena
}
\description{
Un \code{SpatialPolygonsDataFrame} del paquete \code{sp}
    que contiene poligonos usados para la ciudad de Cartagena sub-dividida
    en barrios.
}
\keyword{datasets}
