% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classIntervals.R
\name{classIntervals}
\alias{classIntervals}
\title{Exported function from classInt}
\usage{
classIntervals(
  var,
  n,
  style = "quantile",
  rtimes = 3,
  ...,
  intervalClosure = c("left", "right"),
  dataPrecision = NULL,
  warnSmallN = TRUE,
  warnLargeN = TRUE,
  largeN = 3000L,
  samp_prop = 0.1,
  gr = c("[", "]")
)
}
\arguments{
\item{var}{Variable a mapear a intervalos}

\item{n}{Cantidad de intervalos}

\item{style}{Tipo de intervalo}

\item{rtimes}{Ver documentacion de \code{classInt}}

\item{...}{argumentos adicionales}

\item{intervalClosure}{Donde cierra el intervalo}

\item{dataPrecision}{Ver documentacion de \code{classInt}}

\item{warnSmallN}{Ver documentacion de \code{classInt}}

\item{warnLargeN}{Ver documentacion de \code{classInt}}

\item{largeN}{Ver documentacion de \code{classInt}}

\item{samp_prop}{Ver documentacion de \code{classInt}}

\item{gr}{Ver documentacion de \code{classInt}}
}
\description{
Exported function from classInt
}
