% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{clsf_metrics}
\alias{clsf_metrics}
\title{Medidas de rendimiento para modelos de clasificación}
\usage{
clsf_metrics(fit, split_out, objective = "validation", metrics_out = "all")
}
\arguments{
\item{fit}{El modelos de clasificación ajustado}

\item{split_out}{La partición del conjunto de datos en
entrenamiento/validación/prueba}

\item{objective}{Seleccionar entrenamiento "train" o validación "validation"}

\item{metrics_out}{Vector de caracteres. Incluye los nombres de las
métricas a incluir en la salida
\describe{
    \item{all}{Todas las métricas de clasificación.}
    \item{accuracy}{La **accuracy** general.}
    \item{precision}{La presición binaría o combinada (multi-class)}
    \item{micro_precision}{La precisión binaría para clases individuales.
    Retorna un vector.}
    \item{recall}{La exhaustividad binaría o combinada (multi-class)}
    \item{micro_recall}{La exhaustividad binaría para clases individuales.
    Retorna un vector.}
    \item{macro_f1}{La medida F1 macro obtenida con las medidas macro de
    precisión y exhaustividad.}
    \item{micro_f1}{La medida F1 micro obtenida de las medidas micro de
    precisión y exhaustividad. Retorna un vector.}
    \item{kappa}{El coeficiente Kappa de Cohen que mide el grado de concordancía.}
    \item{adjusted_rand_index}{Medida de similaridad ajustada por chance.
    Puede dar valores negativos.}
    \item{confusion_matrix}{La matriz de confusión. En filas se tiene el valor predecido, en columnas los observados}
}}
}
\value{
Una lista con las métricas seleccionadas o solo el número
    correspondiente a la medida única elegida.
}
\description{
Para modelos de clasificación binaría o múltiple se procesan
    algunas de las principales métricas obtenidas en entrenamiento y
    validación. Las matrices de \code{split_out} deben tener el formato
    requerido por el modelo (escalado, solo numéricas, etc) que se
    solicitaron durante el ajuste del modelo.

    TODO: Ejemplos
}
