% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsf_models.R
\encoding{UTF-8}
\name{clsf_models}
\alias{clsf_models}
\title{Modelos de clasificación - Mapa de datos}
\usage{
clsf_models(
  split_out,
  models = c("multinom_nnet", "multinom_1hl_nnet", "naive_bayes", "svm",
    "decision_tree", "random_forest", "nn_1hl", "nn_2hl"),
  only_best = FALSE,
  save_obj = FALSE,
  ...
)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.
Las columnas de factores deben estar en su correspondiente formato.}

\item{models}{El listado de modelos a evaluar/retornar}

\item{only_best}{Lógico. Retornar solamente el modelo con la mayor
accuracy en validación}

\item{save_obj}{Lógico. Guardar el objeto retornado en disco duro. Se
seleccionado el directorio de trabajo por defecto}

\item{...}{argumentos adicionales pasados a la función principal o a los modelos}
}
\value{
Una lista que contiene los modelos ajustados y las metrícas por modelo
}
\description{
Prototipo de función principal que integra parte de la pipeline
    de modelos de clasificación: selección de modelo, ajuste de modelo y
    reporte de métricas principales. Actualmente la función es capaz de
    procesar los siguientes modelos con sus respectivas especificaciones:
    \describe{
       \item{Multinomial:}{Usando \code{nnet} sin o con una capa oculta}
       \item{Naive Baye:}{Con \code{e1071}}
       \item{SVM}{Con \code{e1071}. Acepta tuning}
       \item{Decision tree}{Con \code{rpart} y pruning automaticó}
       \item{Radom Forest}{Mediante \code{randomForest}}
       \item{NN:}{Redes neuronales con \code{keras} de una o dos capas ocultas}
    }
    Las métricas reportadas son las descritas en \code{clsf_metrics}
}
