% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod_dane.R
\encoding{UTF-8}
\name{cod_dane}
\alias{cod_dane}
\title{Desagregar los códigos DANE en departamento y municipio.}
\usage{
cod_dane(db_obj, cod_col, out_names = c("DPTO_CCDGO", "MPIO_CCDGO"))
}
\arguments{
\item{db_obj}{La base de datos en formato \code{data.frame} o
\code{data.table}}

\item{cod_col}{El nombre que tiene la columna con los códigos}

\item{out_names}{Un vector de dos entradas que contiene el nombre
de las columnas a crear. Primer argumento corresponde a
departamento y segundo a municipio.}
}
\value{
El mismo \code{db_obj} con las columnas agregadas.
}
\description{
Función auxiliar cuyo fin es desglosar los códigos
    DANE municipales completos de bases de datos geo-estadísticas
    en dos columnas de \code{strings} que contienen el código
    del departamento y del municipio (3 dígitos).
}
\examples{
\donttest{
extorsiones <- cod_dane(crim_col_extorsion, 'codigo_dane')
}
}
