% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{col_coor_dept}
\alias{col_coor_dept}
\title{Coordenadas centrales de cada departamento.}
\format{
An object of class \code{data.frame} with 33 rows and 4 columns.
}
\source{
Elaboración propia.
}
\usage{
col_coor_dept
}
\description{
Coordenadas centrales de cada departamento.
}
\keyword{datasets}
