% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{col_dept}
\alias{col_dept}
\title{Departamento de Colombia}
\format{
TODO: Revisar el archivo y depurar.
}
\source{
@source [DANE](https://geoportal.dane.gov.co/servicios/descarga-y-metadatos/descarga-mgn-marco-geoestadistico-nacional/).
}
\usage{
col_dept
}
\description{
Agregación a nivel departamental para los municipios de
    \code{\link{col_mpio}}.
}
\keyword{datasets}
