% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{col_mpio}
\alias{col_mpio}
\title{Información espacial de Colombia.}
\format{
TODO: Revisar el archivo y depurar.
}
\source{
[DANE](https://geoportal.dane.gov.co/servicios/descarga-y-metadatos/descarga-mgn-marco-geoestadistico-nacional/).
}
\usage{
col_mpio
}
\description{
Un \code{SpatialPolygonsDataFrame} del paquete \code{sp}
    que contiene polÍgonos usados en Colombia a nivel de municipio,
    agregables a nivel departamento.
}
\keyword{datasets}
