% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_shinyMobile.R
\name{create_app_ui}
\alias{create_app_ui}
\title{Exported function from shinyMobile}
\usage{
create_app_ui(iframe, device, color, landscape)
}
\arguments{
\item{iframe}{Ver documentacion de \code{shinyMobile}}

\item{device}{Ver documentacion de \code{shinyMobile}}

\item{color}{Ver documentacion de \code{shinyMobile}}

\item{landscape}{Ver documentacion de \code{shinyMobile}}
}
\description{
Exported function from shinyMobile
}
