% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_col_crim.R
\docType{data}
\encoding{UTF-8}
\name{crimcol}
\alias{crimcol}
\alias{crim_col_extorsion}
\alias{crim_col_homicidios}
\alias{crim_col_hurto_1}
\alias{crim_col_hurto_2}
\alias{crim_col_hurto_3}
\alias{crim_col_lesiones}
\alias{crim_col_secuestro}
\alias{crim_col_sexual}
\alias{crim_col_terrorismo}
\alias{crim_col_vio_intra}
\title{Reportes de delitos agregados a nivel de municipio.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 51123 rows and 6 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 174468 rows and 7 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 288104 rows and 7 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 869414 rows and 7 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 433754 rows and 7 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 951909 rows and 7 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 5007 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 199032 rows and 6 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 2198 rows and 7 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 458914 rows and 6 columns.
}
\source{
[Datos abiertos - Colombia](https://www.datos.gov.co/browse?q=reporte%20delitos%20dijin&sortBy=relevance)
}
\usage{
crim_col_extorsion

crim_col_homicidios

crim_col_hurto_1

crim_col_hurto_2

crim_col_hurto_3

crim_col_lesiones

crim_col_secuestro

crim_col_sexual

crim_col_terrorismo

crim_col_vio_intra
}
\description{
Reportes de delitos agregados a nivel de municipio.
}
\keyword{Colombia,}
\keyword{Delitos}
\keyword{datasets}
