% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crime_london_map.R
\encoding{UTF-8}
\name{crime_london_map}
\alias{crime_london_map}
\title{Dibuja un mapa con la localización seleccionada}
\usage{
crime_london_map(map, location, date = NULL, top_n = 5, col_palette = NULL)
}
\arguments{
\item{map}{Un mapa de Londres del paquete \code{leaflet}, creado con
la función \code{\link{draw_city}}. Dejar Nulo para crear un mapa
automatíco usando el conjunto \code{\link{lon_poly_ward_a}}
agregado por *BOROUGH*.}

\item{location}{El nombre del distrito o *ward* a usar. Dejar nulo para
gráficar todo Londres.}

\item{date}{Año y mes donde se registran los hechos, usar el formato
de fechas numéricas 'AAAA-MM' (usar ceros a izquierda de ser
necesario). Consultar fechas disponibles en
\code{available_dates}.}

\item{top_n}{La cantidad de crímenes a reportar. Selecciona las *n*
categorías más frecuentes.}

\item{col_palette}{Una paleta de colores de longitud igual a *n*. Los
colores deben darse en formato hexadecimal o RGB.}
}
\value{
Un mapa de vectores con adición de marcadores según crímenes.
}
\description{
Función creada con el fin de ser ejecutada sin el uso de Shiny.
    Representa una funcionalidad aproximada de la app crime_visor_london donde
    se solicita un mapa ya cargado, una localización y una fecha (puede ser nula).

    Argumentos adicionales pueden ser entregados.
}
\examples{
\donttest{
crime_london_map(NULL, location = 'City of London', date = '2020-01')
}
}
