% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lon_crim.R
\docType{data}
\encoding{UTF-8}
\name{crimlon}
\alias{crimlon}
\alias{lon_violencia_ofensas_sexuales}
\alias{lon_robo_propiedades}
\alias{lon_posesion_armas}
\alias{lon_otros_robos}
\alias{lon_otros_crimenes}
\alias{lon_orden_publico}
\alias{lon_incendios_danos_criminales}
\alias{lon_hurto_personas}
\alias{lon_hurto_comercio}
\alias{lon_hurto_bicicletas}
\alias{lon_drogas}
\alias{lon_crimenes_vehiculos}
\alias{lon_comportamiento_antisocial}
\alias{lon_asalto}
\title{Reportes de delitos granulados en Londres.}
\format{
An object of class \code{data.frame} with 693417 rows and 8 columns.

An object of class \code{data.frame} with 217824 rows and 8 columns.

An object of class \code{data.frame} with 19778 rows and 8 columns.

An object of class \code{data.frame} with 328371 rows and 8 columns.

An object of class \code{data.frame} with 31322 rows and 8 columns.

An object of class \code{data.frame} with 162879 rows and 8 columns.

An object of class \code{data.frame} with 162403 rows and 8 columns.

An object of class \code{data.frame} with 142951 rows and 8 columns.

An object of class \code{data.frame} with 124638 rows and 8 columns.

An object of class \code{data.frame} with 64777 rows and 8 columns.

An object of class \code{data.frame} with 132352 rows and 8 columns.

An object of class \code{data.frame} with 335168 rows and 8 columns.

An object of class \code{data.frame} with 896157 rows and 8 columns.

An object of class \code{data.frame} with 97970 rows and 8 columns.
}
\source{
[UK Police](https://data.police.uk/)
}
\usage{
lon_violencia_ofensas_sexuales

lon_robo_propiedades

lon_posesion_armas

lon_otros_robos

lon_otros_crimenes

lon_orden_publico

lon_incendios_danos_criminales

lon_hurto_personas

lon_hurto_comercio

lon_hurto_bicicletas

lon_drogas

lon_crimenes_vehiculos

lon_comportamiento_antisocial

lon_asalto
}
\description{
Reportes de delitos granulados en Londres.
}
\keyword{Colombia,}
\keyword{Delitos}
\keyword{datasets}
