% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_maps.R
\encoding{UTF-8}
\name{ctab_corplot}
\alias{ctab_corplot}
\title{Gráfico de correlación sobre datos en formatos wide}
\usage{
ctab_corplot(
  x,
  selection = NULL,
  n_show = NULL,
  cor_method = c("kendall", "pearson", "spearman"),
  ...
)
}
\arguments{
\item{x}{Un \code{data.frame} en formato \code{wide}}

\item{selection}{Carácter. El nombre usado para filtrar la visualización}

\item{n_show}{Entero positivo. Cantidad de variables a mostrar más la
seleccionada}

\item{cor_method}{Carácter. Método usado para el cálculo de la matriz de
correlaciones}

\item{...}{Argumentos adicionales}
}
\value{
Una gráfica de correlación
}
\description{
Se usa un \code{data.frame} en formato \code{wide} de datos
    numéricos clasificados por columnas (variables) y filas
    (individuos/tiempos). El gráfico es procesado usando \code{corrplot}
}
