% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_maps.R
\encoding{UTF-8}
\name{ctab_lag}
\alias{ctab_lag}
\title{Aplicar rezagos a tablas en formato wide}
\usage{
ctab_lag(x, selection, n_lags = 1)
}
\arguments{
\item{x}{El \code{data.frame} a tratar}

\item{selection}{El nombre de la columna a usar como referencia. Sobre ella
no se aplican los rezagos}

\item{n_lags}{Entero. La cantidad de rezagos a usar en las variables}
}
\value{
El conjunto de datos con las variables rezagadas. Se eliminan las
    filas donde las variables rezagadas no tienen observaciones
}
\description{
En principio la función se dirige a \code{data.frames}
    provenientes del proceso \code{ctab_period} aunque funciona para
    cualquier tipo de conjunto que cumpla con poseer un formato wide
    y no tenga otras columnas diferentes a las expandidas
}
\examples{
# Usamos la base de datos de Bogota, filtrando por robo de celulares
x <- siedco_wc[Delito == "HURTO DE CELULARES",,]
periodo <- c("2019-03-01","2020-12-30")
ejemplo <- ctab_period(x, date_range = periodo, div_name = "NOM_DIV1",
var_agg = "Numero Hechos")
# Rezago de dos meses para PUENTE ARANDA
ctab_lag(ejemplo, "PUENTE ARANDA", 2)
}
