% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_maps.R
\encoding{UTF-8}
\name{ctab_period}
\alias{ctab_period}
\title{Creación de tablas en formato wide para mapas de delito}
\usage{
ctab_period(x, date_range, div_name, var_agg = "N", ...)
}
\arguments{
\item{x}{El conjunto a procesar. Se admite que los datos ya hayan sido
agrupados y agregados previamente, pero se recomienda incluir
el parámetro adicional \code{pre_pro = FALSE}}

\item{date_range}{Un vector con la fecha inicial y final. Se sugiere usar
un formato como "aaaa-mm-dd". El orden de las fechas debe mantenerse}

\item{div_name}{El nombre de la columna que contiene el nombre de las
divisiones espaciales, o sobre la cual se consolidan los grupos (más
las fechas)}

\item{var_agg}{El nombre de la columna que contienen los valores a ser
agregados}

\item{...}{Argumentos adicionales. La columna de fechas se asume con el
nombre "FECHA" (sin distinción de mayúsculas), en caso contrario
incluir el argumento \code{date_col_name} con el nombre correspondiente.
Se realiza una agregación de las fechas por meses, si se desea usar
otra unidad como semanas o días se debe pasar la función que haga este
procedimiento con el argumento \code{date_agg_fun}}
}
\value{
Un \code{data.frame} con formato wide
}
\description{
La intención de esta función es asistir en la representación
    de los conjuntos de datos de delitos que originalmente son presentados
    como número de ocurrencias por fecha y unidad territorial
    (e.g. Localidad, UPZ), agregándolos en los grupos inducidos por las
    variables temporales/espaciales de acuerdo a un patrón como
    meses/división territorial uno o similares. Puede ser usado con
    otro tipo de datos pero se necesita pasar parámetros adicionales
}
\examples{
# Usamos la base de datos de Bogota, filtrando por robo de celulares
x <- siedco_wc[Delito == "HURTO DE CELULARES",,]
periodo <- c("2019-03-01","2020-12-30")
ctab_period(x, date_range = periodo, div_name = "NOM_DIV1",
var_agg = "Numero Hechos")
}
