% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2ctgr.R
\encoding{UTF-8}
\name{ctgr2nmd}
\alias{ctgr2nmd}
\title{Asignar nombres a las categorías ordenadas (intervalos)}
\usage{
ctgr2nmd(intrvls, names_intr)
}
\arguments{
\item{intrvls}{Un objeto devuelto por la función \code{num2ctgr} o
\code{add2ctgr}}

\item{names_intr}{Los nombres (ordenados) a ser asignados}
}
\value{
Un vector de factores con niveles ordenados (menor a mayor)
}
\description{
Asignar nombres a las categorías ordenadas (intervalos)
}
\examples{
example_vect <- c(-1,-1,-10,1,0,2,3,10,5,11,12,15,20,-4)
example_int <- num2ctgr(x = example_vect, type = "jenks", n = 2)
ctgr2nmd(example_int, c("Bajo", "Alto"))
}
