% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_models_city.R
\encoding{UTF-8}
\name{custom_models_city}
\alias{custom_models_city}
\title{Evaluación de modelos creados por el usuario para delitos seleccionados.}
\usage{
custom_models_city(city, crimes, ...)
}
\arguments{
\item{city}{La ciudad a evaluar}

\item{crimes}{Un vector con el nombre de los delitos a consultar.}

\item{...}{Los parámetros con nombres de los modelos corridos. Los parámetros
admitidos (con su nombre) son \code{p_max}, \code{date_mix},
\code{tunning_intercept} y \code{split_prop}. Consultar las documentaciones
correspondientes para más detalles.}
}
\value{
Una lista con el nombre de cada conjunto de datos analizado que
    incluye la salida del mejor modelo. Para más detalles ver la
    documentación respectiva de cada algoritmo de selección.
}
\description{
Primera versión de la rutina conjunta para el reporte completo
    de modelos personalizados para una ciudad de elección y una lista de
    delitos. De momento se debe supervisar que la combinación ciudad+delito
    devuelva el nombre de una base de datos existente, para ello se recomienda
    consultar \code{data} y \code{data_col_crim}.

    En esta versión se evaluán los modelos de \code{lin_mod_selector}
    y \code{ar_mod_selector}.

    TODO: Permitir que dots admita la especificación de modelo(s). Ejemplos.
}
