% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{cvpred_metric}
\alias{cvpred_metric}
\title{Pseudo-coeficiente de variación para la predicción}
\usage{
cvpred_metric(model, split_output)
}
\arguments{
\item{model}{El modelo de regresión ajustado a medir su rendimiento}

\item{split_output}{La partición del conjunto de datos en
entrenamiento/validación/prueba}
}
\value{
El RMSE fuera de muestra del objeto.
}
\description{
Usada para obtener la medida de rendimiento sobre modelos
    definidos por el usuario. La medición se ejecuta sobre el conjunto
    de validación. La rutina asume que X existe y es proporcionada.

    TODO: Ejemplos.
}
