% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{date_split}
\alias{date_split}
\title{Procesamiento de fechas para adicionar características}
\usage{
date_split(x, ...)
}
\arguments{
\item{x}{El vector de fechas en un formato apropiado como "AAAA-MM-DD".}

\item{...}{Una serie de pares de clave/valor donde los últimos son
\code{TRUE} (por defecto) o \code{FALSE}. Las claves corresponden con
los campos que salida de la función, en este caso son:
\describe{
     \item{\code{year}:}{Devolver el año}
     \item{\code{month}:}{Devolver el mes númerico}
     \item{\code{week}:}{Devolver el número de la semana}
     \item{\code{week_day}:}{Devolver el día de la semana}
     \item{\code{day_range}:}{Devolver el rango del día}
}}
}
\value{
Una lista con los nuevos campos.
}
\description{
Acepta un vector de fechas, que pueden ser \code{string},
    \code{Date} o \code{POSIXct}, del cual se extraen alguna de las
    siguientes componentes: año, mes, semana, día de la semana y/o
    rango del día.
}
\examples{
# Crear una secuencia de fechas
X <- seq(as.Date("2020-01-01"), length = 10, by = "1 day")
# Particion completa
date_split(X)
}
