% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{db_crime_london_a}
\alias{db_crime_london_a}
\title{Base de datos de crimen Feb-2018 a Ene-2021.}
\format{
An object of class \code{list} of length 33.
}
\source{
[UK Police](https://data.police.uk/)
}
\usage{
db_crime_london_a
}
\description{
Base de datos de crimen Feb-2018 a Ene-2021.
}
\keyword{datasets}
