% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sp_merge.R
\encoding{UTF-8}
\name{db_sp_merge}
\alias{db_sp_merge}
\title{Join de una base datos y un \code{SpatialPolygonsDataFrame}}
\usage{
db_sp_merge(db_obj, sp_obj, merge_on)
}
\arguments{
\item{db_obj}{Una base de datos en formato \code{data.frame} o
\code{data.table}.}

\item{sp_obj}{Un objeto \code{SpatialPolygonsDataFrame} del paquete
\code{sp}.}

\item{merge_on}{Las columnas sobre las cuales se realiza el join.}
}
\value{
Un \code{SpatialPolygonsDataFrame}
}
\description{
Realiza un join a izquierda del objeto
    \code{SpatialPolygonsDataFrame} con la base de datos suministrada.
}
\examples{
\dontrun{
list_filter <- list(tipo_de_hurto = 'HURTO RESIDENCIAS',
    fecha_hecho = data.table::as.IDate('2020-01-01'))
hurtos <- test2021:::dt_filter_crime(crim_col_hurto_1, list_filter)
hurtos <- cod_dane(hurtos, 'codigo_dane')
colombia_extorsiones <- db_sp_merge(hurtos, col_mpio,
c('DPTO_CCDGO', 'MPIO_CCDGO'))
}
}
