% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_city.R
\encoding{UTF-8}
\name{draw_city}
\alias{draw_city}
\title{Dibujar el mapa de la ciudad seleccionada}
\usage{
draw_city(city_file = lon_poly_ward_a, merge_on = NULL)
}
\arguments{
\item{city_file}{Un objeto de tipo \code{SpatialPolygonDataFrame}.}

\item{merge_on}{En que unidades el mapa debe integrar sus divisiones.
Consultar las columnas del archivo \code{city_file} suministrado
sobre las cuales se agregan los polígonos.}
}
\value{
Un mapa de vectores. Por defecto un mapa de Londres y sus
    *wards*.
}
\description{
Dibujar el mapa de la ciudad seleccionada
}
\examples{
\donttest{
draw_city()
}
}
