% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{fill_long}
\alias{fill_long}
\title{Completar datos panel/longitudinales por el índice de tiempo}
\usage{
fill_long(y, x, date_col, time_delta = "1 day", fill = 0)
}
\arguments{
\item{y}{El vector de valores observados para las series.}

\item{x}{El \code{data.frame} de datos que contiene tanto la información de
fechas como de los factores complementarios para distinguir los
individuos.}

\item{date_col}{El nombre que lleva la columna de fechas.}

\item{time_delta}{El tamaño de los saltos entre observaciones. Debe ser
un formato reconocido para fechas (i.e. unidades temporales)}

\item{fill}{El valor constante con el que completar los datos perdidos.}
}
\value{
La serie \code{y} aumentada para el resto de fechas.
}
\description{
Función creada para completar múltiples series de tiempo donde se
    dispone de una grilla regular de observación pero que en algunos puntos
    las observaciones no estén disponibles. Cada serie de tiempo esta asociada
    a un *individuo* que esta determinado por la agrupación de ciertos factores
    que se definen en el \code{data.frame} pasado como argumento. Se asume
    que todos los individuos son observados en el mismo horizonte de tiempo.

    El valor usado para completar las observaciones perdidas se específica
    como argumento.
}
\examples{
# Creamos una serie de fechas y se toma una fraccion
X_fecha <- seq(as.Date("2020-01-01"), length = 10, by = "1 day")
X_fecha <- X_fecha[c(1,sample(2:9,4),10)]
X <- data.frame(FECHA = rep(X_fecha), LETRA = rep(c("A","B"), each = 6))
Y <- rnorm(12)
# Serie llenada con ceros
fill_long(Y, X, "FECHA")
}
