% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{fill_series}
\alias{fill_series}
\title{Completar series univariadas de intervalos regulares}
\usage{
fill_series(y, dates, time_delta = "1 day", fill = 0)
}
\arguments{
\item{y}{El vector de valores observados para la serie.}

\item{dates}{Los tiempos (fechas) donde fue observada la serie.}

\item{time_delta}{El tamaño de los saltos entre observaciones. Debe ser
un formato reconocido para fechas (i.e. unidades temporales)}

\item{fill}{El valor constante con el que completar los datos perdidos.}
}
\value{
La serie \code{y} aumentada para el resto de fechas.
}
\description{
Función creada para completar series univariadas donde se
    disponga de una grilla regular de observación pero en algunos puntos
    las observaciones no estén disponibles. El valor usado para completar
    las observaciones perdidas se específica como argumento.
}
\examples{
# Creamos una serie de fechas y se toma una fraccion
X_fecha <- seq(as.Date("2020-01-01"), length = 10, by = "1 day")
X_fecha <- X_fecha[c(1,sample(2:9,4),10)]
Y <- rnorm(6)
# Serie llenada con ceros
fill_series(Y,X_fecha)
}
