% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{filter_crime}
\alias{filter_crime}
\title{Filtro crimen/fecha - POR REMOVER}
\usage{
filter_crime(x, crime_s, date_s, crim_col, date_col)
}
\arguments{
\item{x}{Datos a filtrar.}

\item{crime_s}{Crimen (string).}

\item{date_s}{Fecha (string).}

\item{crim_col}{Nombre columna con crimenes.}

\item{date_col}{Nombre columna con fechas.}
}
\description{
Filtro crimen/fecha - POR REMOVER
}
