% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cartogram.R
\encoding{UTF-8}
\name{get_cartogram}
\alias{get_cartogram}
\title{Crear un cartograma a partir de objecto \code{sp}/\code{sf}.}
\usage{
get_cartogram(x, var_weight, crs, type = "cont", itermax = 10)
}
\arguments{
\item{x}{Un objeto \code{SpatialPolygonsDataFrame} o \code{sf}.}

\item{var_weight}{El nombre de la columna de la variable numérica sobre
la cual se asignan pesos.}

\item{crs}{El número de la proyección a usar.}

\item{type}{El tipo de cartograma, puede ser: cont o dorling.}

\item{itermax}{La cantidad máxima de veces que el algoritmo
itera.}
}
\value{
Una sf que contiene el cartograma calculado.
}
\description{
Se suministra un \code{SpatialPolygonsDataFrame} o un
    \code{sf} que contengan una variable sobre la cual se pueda
    asignar pesos a las propias geometrías o a círculos dentro de
    estas, cuyos tamaños reflejen los valores de dicha variable, y
    luego la función retorna un gráfico  en \code{ggplot2}
    que muestra el respectivo cartograma.

    Los dos tipos de cartogramas posibles son \code{cont} o continuo
    el cual deforma las geometrías en función del peso, haciendo más
    pequeños los polígonos donde este sea menor y viceversa. Por otro
    lado se tiene el cartograma \code{dorling} que crea los círculos
    ya mencionados y los posiciona cerca a su lugar de origen, pero
    sin que alguno se solape.
}
\examples{
\dontrun{
# Dibujar los departamentos de Colombia segun su poblacion
# Usar el CRS de Colombia - 3118
get_cartogram(col_dept, 'TOTPOP', 3118)
}
}
