% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{get_centroid}
\alias{get_centroid}
\title{Obtención de centroide para visualización}
\usage{
get_centroid(unit, data = lon_poly_ward_a, geo_div = "BOROUGH")
}
\arguments{
\item{unit}{Unidad de división geo-política.}

\item{data}{Conjunto de datos espaciales:
\code{\link{lon_poly_ward_a}} o \code{\link{lon_poly_ward_b}}.}

\item{geo_div}{Nombre de la columna usada para crear las divisiones.
Ver posibles valores en \code{\link{lon_poly_ward_a}} o
\code{\link{lon_poly_ward_b}}.}
}
\value{
Centroide
}
\description{
Obtención de centroide para visualización
}
