% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cl.R
\encoding{UTF-8}
\name{get_cl}
\alias{get_cl}
\title{Obtener líneas de contorno para gráficas}
\usage{
get_cl(df_coords, grid_size = c(50, 50), band_width = NULL)
}
\arguments{
\item{df_coords}{Un \code{data.frame} o \code{tibble} con
columnas de latitud o longitud. Los nombres de ellas deben
poseer un nombre que asemeje la descripción dada.}

\item{grid_size}{Un vector que contiene el número de puntos
equidistantes en cada dirección sobre los que se estima
la densidad.}

\item{band_width}{Un vector con dos números positivos que determina
la longitud de banda para el estimador del kernel.}
}
\value{
Una lista que contiene los siguiente elementos:
\describe{
    \item{sp_obj}{Un objeto \code{SpatialPolygons}.}
    \item{cl}{Las líneas de contorno calculadas.}
    \item{cl_lev}{Los niveles de las líneas de contorno,
    usados para asignar colores en los mapas.}
    \item{cl_nlev}{El número de niveles de las líneas
    de contorno, usado para asignar colores en los mapas.}
}
}
\description{
Creación de líneas de contorno a partir de kernels
    normales bivariados estimados sobre un conjunto de datos
    de coordenadas espaciales y evaluado sobre una grilla.
}
\examples{
\dontrun{
crime_1000 <- test2021:::list_to_df(db_crime_london_a)[1:1000,]
get_cl(crime_1000)
}
}
