% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{get_crim_cat}
\alias{get_crim_cat}
\title{Transformación de categorías en código a valores significativos}
\usage{
get_crim_cat(x, lang = "ESP")
}
\arguments{
\item{x}{Base de datos de crímenes}

\item{lang}{Lenguaje usada en tabla}
}
\value{
Categorías traducidas
}
\description{
Transformación de categorías en código a valores significativos
}
