% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crime_london.R
\encoding{UTF-8}
\name{get_crime_london}
\alias{get_crime_london}
\title{Obtener datos de crimen por fecha y localización (Londres)}
\usage{
get_crime_london(
  data = lon_poly_ward_a,
  geo_div = "BOROUGH",
  date = NULL,
  lang = "ESP",
  static = FALSE
)
}
\arguments{
\item{data}{Un objeto \code{SpatialPolygonDataFrame} correspondiente
a Londres. Conjuntos disponibles:
\code{\link{lon_poly_ward_a}} o \code{\link{lon_poly_ward_b}}.}

\item{geo_div}{Nombre de la columna usada para crear las divisiones.
Ver posibles valores en \code{\link{lon_poly_ward_a}} o
\code{\link{lon_poly_ward_b}}.}

\item{date}{Año y mes donde se registran los hechos, usar el formato
de fechas numéricas 'AAAA-MM' (usar ceros a izquierda de ser
necesario). Consultar fechas disponibles en
\code{available_dates}.}

\item{lang}{Lenguaje en que se reportan los crímenes: inglés 'ENG'
o español 'ESP'.}

\item{static}{Bolean. Usar o no el conjunto pre-cargado. Solo
disponible para \code{\link{lon_poly_ward_b}} y nivel distritos.}
}
\value{
Un \code{tibble} que integra la información de crímenes.
        [Ver documentación](https://data.police.uk/docs/)
}
\description{
Método del paquete que llama la función \code{ukpolice::ukc_crime_poly}
    sobre cada una de las divisiones establecidas sobre el mapa, ya sean distritos o
    *wards*. Los datos son aproximados por anonimización de las fuentes, según la
    documentación de la [API base](https://data.police.uk/docs/) proporcionada por
    el departamento de policía de Reino Unido.

    Para que esta función pueda correr sin problema alguno, se requiere usar alguno
    de las base de datos pre-cargadas y llamar una de las divisiones especificadas.
    No es posible obtener toda la información de Londres dentro de un solo REQUEST
    por parte de la API (Error 503). Es posible que en algunos casos no se reporte
    información para algún año-mes dado
}
\examples{
\dontrun{
get_crime_london()
}
}
