% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crime_london.R
\encoding{UTF-8}
\name{get_crime_london_unit}
\alias{get_crime_london_unit}
\title{Obtener datos de crimen por fecha y localización (Londres) - Unidad}
\usage{
get_crime_london_unit(
  unit,
  data = lon_poly_ward_a,
  geo_div = "BOROUGH",
  date = NULL,
  lang = "ESP",
  static = FALSE
)
}
\arguments{
\item{unit}{Nombre de la localización. Ver posibles valores en
\code{\link{get_london_boroughs}} o
\code{\link{get_london_districts}}.}

\item{data}{Un objeto \code{SpatialPolygonDataFrame} correspondiente
a Londres. Conjuntos disponibles:
\code{\link{lon_poly_ward_a}} o \code{\link{lon_poly_ward_b}}.}

\item{geo_div}{Nombre de la columna usada para crear las divisiones.
Ver posibles valores en \code{\link{lon_poly_ward_a}} o
\code{\link{lon_poly_ward_b}}.}

\item{date}{Año y mes donde se registran los hechos, usar el formato
de fechas numéricas 'AAAA-MM' (usar ceros a izquierda de ser
necesario). Consultar fechas disponibles en
\code{available_dates}.}

\item{lang}{Lenguaje en que se reportan los crímenes: inglés 'ENG'
o español 'ESP'.}

\item{static}{Bolean. Usar o no el conjunto pre-cargado. Solo
disponible para \code{\link{lon_poly_ward_b}} y nivel distritos.}
}
\value{
Un \code{tibble} que integra la información de crímenes.
        [Ver documentación](https://data.police.uk/docs/)
}
\description{
Método del paquete que llama la función \code{ukpolice::ukc_crime_poly}
    sobre **UNA** de las divisiones establecidas sobre el mapa, ya sean distritos o
    *wards*. Los datos son aproximados por anonimización de las fuentes, según la
    documentación de la [API base](https://data.police.uk/docs/) proporcionada por
    el departamento de policía de Reino Unido.
}
\examples{
\donttest{
get_crime_london_unit('City of London')
}
}
