% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{get_london_boroughs}
\alias{get_london_boroughs}
\title{Devuelve un listado de los *boroughs* (Londres)
    disponibles}
\usage{
get_london_boroughs()
}
\value{
Listado de *boroughs* de Londres.
}
\description{
Devuelve un listado de los *boroughs* (Londres)
    disponibles
}
