% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{get_london_districts}
\alias{get_london_districts}
\title{Devuelve un listado de los  distritos (Londres)
    disponibles}
\usage{
get_london_districts()
}
\value{
Listado de distritos de Londres.
}
\description{
Devuelve un listado de los  distritos (Londres)
    disponibles
}
