% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_bins.R
\encoding{UTF-8}
\name{hex_bins}
\alias{hex_bins}
\title{Creación de polígonos hexagonales para mapas}
\usage{
hex_bins(
  db_pnts,
  crs,
  n_hex = 500,
  buffer = 0.05,
  agg_col = NULL,
  agg_fun = "count"
)
}
\arguments{
\item{db_pnts}{Un \code{data.frame} o \code{data.table} que contiene
dos columnas de latitud o longitud, cuyos nombres deben reflejar
esto. Además, de usarse \code{agg_col}, esta debe de estar incluida
acá.}

\item{crs}{El número del sistema de proyección utilizada. Por defecto
se utiliza la EPSG:3118.}

\item{n_hex}{La cantidad de hexágonos que tiene la grilla.}

\item{buffer}{Un buffer a la distancia de al casco convexo usado para
crear la grilla hexagonal. Por defecto es 5\%.}

\item{agg_col}{El nombre de la columna sobre la cual se agrega. Dejar
nulo si se usa \code{agg_fun} como \code{count}.}

\item{agg_fun}{El nombre de la función que agrega o una expresión que
se pueda pasar a \code{R}.}
}
\value{
Una grilla hexagonal contenida en una clase
    \code{SpatialPolygonsDataFrame}.
}
\description{
Transforma una base de datos de puntos geo-localizados
    (latitud-longitud) en una grilla de hexágonos con información
    adjunta, ideal para la creación de gráficas y mapas que soporten
    los objetos tipo \code{SpatialPolygonsDataFrame}.
}
\examples{
\dontrun{
# Base de datos
db_1 <- test2021:::list_to_df(db_crime_london_a)
db_1 <- data.table::as.data.table(db_1)
# Filtrada
db_2 <- db_1[ category == 'Drogas' & month == '2019-01']
hex_drogras <- hex_bins(db_2, crs = 27700)
}
}
