% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_mod_selector.R
\encoding{UTF-8}
\name{lin_mod_selector}
\alias{lin_mod_selector}
\title{Ajuste de un modelo de regresión múltiple con validación por RMSE.}
\usage{
lin_mod_selector(
  split_out,
  step_selection = TRUE,
  p_selection = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.}

\item{step_selection}{¿Se debe probar subconjuntos de variables?}

\item{p_selection}{¿Se debe eliminar las variables con p-valores no
significativos?}

\item{quiet}{Callar las salidas en consola de los modelos.}
}
\value{
Una lista que incluye el mejor modelo y los RMSE en entrenamiento
    y validación (un pase)
}
\description{
Modelo personalizado que compara un ajuste de un modelo
    contra uno reducido, seleccionado por procedimientos automáticos.
    Se entregan los RMSE de entrenamiento y validación, junto al
    modelo seleccionado.

    No se realizan múltiples pases de validación.

    TODO: Ejemplos y múltiples pases de validación.
    TODO: Es necesario corregir la asignacion global de train...
}
