% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{linmod_metrics}
\alias{linmod_metrics}
\title{Medidas de rendimiento para modelos de regresión}
\usage{
linmod_metrics(model, split_output, objective = "validation", ...)
}
\arguments{
\item{model}{El modelo de regresión ajustado a medir su rendimiento}

\item{split_output}{La partición del conjunto de datos en
entrenamiento/validación/prueba}

\item{objective}{Seleccionar entrenamiento "train" o validación "validation"}

\item{...}{Los nombres de las medidas adicionales, entre las que se
encuentran:
\describe{
    \item{\code{mean}}{media de los residuos.}
    \item{\code{mean_observed}}{media de la variable respuesta.}
    \item{\code{mae}}{error absoluto medio de los residuos.}
    \item{\code{cve_rmse}}{cociente entre \code{rmse} y \code{mean}.}
    \item{\code{sd_observed}}{desviación estándar de la variable respuesta.}
    \item{\code{rmse_scaled}}{RMSE obtenido con residuales estandarizados por la suma absoluta de las diferencias de las observaciones.}
}}
}
\value{
Una lista de medidas de rendimiento dentro de muestra o fuera,
    según la elección del usuario
}
\description{
Usada para obtener diferentes medidas para evaluación de
    rendimiento sobre modelos definidos por el usuario. La medición
    se realiza sobre el conjunto de entrenamiento o validación.
    Se asume que X existe y es proporcionada.

    Por defecto entrega el RMSE.

    TODO: Ejemplos.
}
