% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{list_to_df}
\alias{list_to_df}
\title{Convertir lista a \code{data.frame}}
\usage{
list_to_df(z, col_name = "DISTRICT")
}
\arguments{
\item{z}{Una lista de \code{data.frames}}

\item{col_name}{Nombre que se asigna a la columna creada.}
}
\description{
El nombre de los elementos se convierte en una columna.
}
