% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{lon_poly_ward_b}
\alias{lon_poly_ward_b}
\title{Información espacial de Londres (2018)}
\format{
TODO: Revisar el archivo y depurar.
}
\source{
[Greater London Authority](https://data.london.gov.uk/dataset/statistical-gis-boundary-files-london) (GLA)
}
\usage{
lon_poly_ward_b
}
\description{
Un \code{SpatialPolygonsDataFrame} del paquete \code{sp}
    que contiene polÍgonos usados para el distrito de Londres dividido
    en *wards* (de acuerdo a la GLA).
}
\keyword{datasets}
