% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{londres}
\alias{londres}
\title{Información espacial de Gran Londrés}
\format{
TODO: Revisar el archivo.
}
\usage{
londres
}
\description{
Un \code{SpatialPolygonsDataFrame} del paquete \code{sp}
    que contiene poligonos usados para Gran Londrés con sus diviciones
    en distritos y barrios. Esta versión es compatible con las app de
    ciudades.
}
\keyword{datasets}
