% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{look_colombia_crime}
\alias{look_colombia_crime}
\title{Tabla de valores para codificación para Colombia.}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\source{
Elaboración propia.
}
\usage{
look_colombia_crime
}
\description{
Tabla de valores para codificación para Colombia.
}
\keyword{datasets}
