% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{look_crimes_cat}
\alias{look_crimes_cat}
\title{Tabla de valores para codificación de crimenes de Londres.}
\format{
An object of class \code{data.frame} with 15 rows and 3 columns.
}
\source{
Elaboración propia.
}
\usage{
look_crimes_cat
}
\description{
Tabla de valores para codificación de crimenes de Londres.
}
\keyword{datasets}
