% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_col_crim.R
\docType{data}
\encoding{UTF-8}
\name{med_hurto_bicicletas}
\alias{med_hurto_bicicletas}
\title{Reportes simulados de robos de bicicletas en Medellín}
\format{
An object of class \code{data.frame} with 12330 rows and 7 columns.
}
\source{
Elaboración propia
}
\usage{
med_hurto_bicicletas
}
\description{
Reportes simulados de robos de bicicletas en Medellín
}
\keyword{datasets}
