% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{medellin}
\alias{medellin}
\title{Información espacial de Medellín}
\format{
TODO: Revisar el archivo.
}
\source{
[Alcaldía de Medellín - Open Data](https://geomedellin-m-medellin.opendata.arcgis.com/datasets/barrio-vereda)
}
\usage{
medellin
}
\description{
Un \code{SpatialPolygonsDataFrame} del paquete \code{sp}
    que contiene poligonos usados para la ciudad de Medellín sub-dividida
    en barrios y comunas.
}
\keyword{datasets}
