% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{movilidad_abril_bog}
\alias{movilidad_abril_bog}
\title{Datos de movilidad agregados para Abril de 2019.}
\format{
An object of class \code{sf} (inherits from \code{grouped_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4365 rows and 7 columns.
}
\source{
Secretaria de movilidad de Bogotá - SIMUR
}
\usage{
movilidad_abril_bog
}
\description{
Datos de movilidad agregados para Abril de 2019.
}
\keyword{datasets}
