% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{movilidad_marzo_bog}
\alias{movilidad_marzo_bog}
\title{Datos de movilidad agregados para Marzo de 2021.}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11198 rows and 17 columns.
}
\source{
Secretaria de movilidad de Bogotá - SIMUR
}
\usage{
movilidad_marzo_bog
}
\description{
Datos de movilidad agregados para Marzo de 2021.
}
\keyword{datasets}
