% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{multi_model_output}
\alias{multi_model_output}
\title{Resumenes para modelos agrupados}
\usage{
multi_model_output(model_list, n_model = 0)
}
\arguments{
\item{model_list}{Lista de modelos. Ver \code{arima_mod_selector}.}

\item{n_model}{El número del modelo a retornar. Usar 0 para devolver todos.}
}
\value{
Imprime en consola el resumen del modelo ajustado.
}
\description{
Resumenes para modelos agrupados
}
