% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_mod_selector.R
\encoding{UTF-8}
\name{multinom_mod_selector}
\alias{multinom_mod_selector}
\title{Ajuste de modelos multinomiales log-lineales con RN}
\usage{
multinom_mod_selector(split_out, scaled = FALSE, quiet = FALSE, size = 0, ...)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.
Las columnas de factores deben estar en su correspondiente formato.}

\item{scaled}{¿Los datos están escalados? De usar falso se realizará
su escalamiento. *Nota:* El escalamiento se realiza usando la
información disponible en entrenamiento.}

\item{quiet}{Lógico. Callar las salidas en consola de los modelos.}

\item{size}{Entero positivo. Determina el número de nodos de la capa oculta.
Elegir cero para usar una NN sin capa oculta.}

\item{...}{argumentos adicionales pasados a \code{multinom} o \code{nnet}}
}
\value{
Una lista que incluye el mejor modelo multinomial junto a la
    **accuracy** reportada en entrenamiento y validación.
}
\description{
Modelos log-lineales multinomiales ajustados usando redes
    neuronales con la rutina \code{multinomial} y \code{nnet}
    del paquete \code{nnet}. Se reporta la **accuracy** en entrenamiento
    y validación.

    Dado que esto es un modelo de clasificación, la respuesta esperada
    debe ser un \code{factor}. Además, por como funciona las redes
    neuronales de \code{nnet} las columnas numéricas deben estar escaladas
    para asegurar rápida convergencia.

    Para seleccionar el número de nodos en la red neuronal se recomienda
    elegir una cantidad entre el número de *inputs* y el de *outputs*.

    *Nota* en versiones futuras se incluirá ajuste por CV para el
    pesos de los parámetros y el número de nodos
    (usando \code{caret}, por ejemplo)

    TODO: Ejemplos
}
