% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_mod_selector.R
\encoding{UTF-8}
\name{nb_mod_selector}
\alias{nb_mod_selector}
\title{Ajuste de clasificador probabilistico bayesiano (Naïve Bayes)}
\usage{
nb_mod_selector(split_out, scaled = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.
Las columnas de factores deben estar en su correspondiente formato.}

\item{scaled}{¿Los datos están escalados? De usar falso se realizará
su escalamiento. *Nota:* El escalamiento se realiza usando la
información disponible en entrenamiento.}

\item{quiet}{Lógico. Callar las salidas en consola de los modelos.}

\item{...}{argumentos adicionales (en desarrollo)}
}
\value{
Una lista que incluye el mejor clasificador probabilistico
    junto a la **accuracy** reportada en entrenamiento y validación.
}
\description{
Mediante la rutina \code{naiveBayes} se obtienen las
    probabilidades posteriores de cada clase en función de los predictores
    independientes suministrados. Se asume gaussianidad para las variables
    continuas. En la salida se reportan las **accuracy** del árbol simplificado
    en entrenamiento y validación.

    *Nota* en versiones futuras se incluirá ajuste por CV
    (usando \code{caret}, por ejemplo)

    TODO: Ejemplos
}
