% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_mod_selector.R
\encoding{UTF-8}
\name{nn_2hl}
\alias{nn_2hl}
\title{Redes neuoronales de dos capas ocultas}
\usage{
nn_2hl(split_out, split_transform = TRUE, early_stop = TRUE, ...)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.
Las columnas de factores deben estar en su correspondiente formato.}

\item{split_transform}{Lógico. Aplicar encoding a los datos de ser necesario}

\item{early_stop}{Lógico. Usar para temprana para mejorar generalización}

\item{...}{argumentos adicionales pasados a Keras}
}
\value{
Un modelo ajustado en Keras
}
\description{
Por defecto usadas para clasificación con salida softmax. Se
    asume que los datos vienen estandarizados (numéricos)
}
\examples{
split_example <- create_split_ts(mtcars$vs, subset(mtcars, select = -vs),
split_prop = c(0.7,0.25,0.05))
split_example$Y$Y_train <- as.factor(split_example$Y$Y_train)
split_example$Y$Y_val <- as.factor(split_example$Y$Y_val)
split_example$Y$Y_test <- as.factor(split_example$Y$Y_test)
modelo <- nn_2hl(split_example)
}
