% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2ctgr.R
\encoding{UTF-8}
\name{num2ctgr}
\alias{num2ctgr}
\title{Creación de categorías (intervalos) a partir de un vector numérico}
\usage{
num2ctgr(x, type = c("quantiles", "jenks", "subjective"), ...)
}
\arguments{
\item{x}{El vector numérico a convertir}

\item{type}{El método usado para crear los intervalos. Ver \code{...}}

\item{...}{Argumentos adicionales que son pasados de acuerdo con \code{type}
\describe{
    \item{probs}{cuando \code{type = "quantiles"} se debe pasar los
    cortes de los cuantiles requeridos (puede omitirse 0 y 1) como
    vector numérico}
    \item{n}{cuando \code{type = "jenks"} se requiere la cantidad
    positiva entera de clases}
    \item{breaks}{cuando \code{type = "subjective"} se debe pasar
    los puntos de cortes para crear los intervalos como vector
    numérico}
}}
}
\value{
Una lista con nombres que contiene los intervalos (\code{intrvls})
    y los cortes usados para crearlos (\code{brks})
}
\description{
Procesa un vector numérico en una serie de intervalos definidos
    por el método seleccionado por el usuario, y según el cual se deberán
    incluir los argumentos adicionales correspondientes. Para convertir
    los intervalos a categorías con nombres se debe aplicar al objeto
    devuelto la función \code{ctgr2nmd}
}
\examples{
example_vect <- c(-1,-1,-10,1,0,2,3,10,5,11,12,15,20,-4)
num2ctgr(x = example_vect, type = "jenks", n = 2)
num2ctgr(x = example_vect, type = "quantiles", probs = c(0.3,0.6))
}
