% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\encoding{UTF-8}
\name{one_hot}
\alias{one_hot}
\title{One-hot encoding para conjuntos de datos mixtos}
\usage{
one_hot(
  dt,
  cols = "auto",
  sparsifyNAs = FALSE,
  naCols = FALSE,
  dropCols = TRUE,
  dropUnusedLevels = FALSE
)
}
\arguments{
\item{dt}{Un \code{data.table} o \code{data.frame}}

\item{cols}{Vector con el nombre de las columnas a ser codificadas. Por
defecto se aplica a todos los factores}

\item{sparsifyNAs}{Lógico. Convertir los NA en 0}

\item{naCols}{Lógico. Crear o no columnas que indiquen presencia de NA
(de existir)}

\item{dropCols}{Lógico. El resultado debe incluir o no las columnas
originales sin codificar}

\item{dropUnusedLevels}{Lógico. Crear o no columnas de solo ceros para
niveles de factores no usados}
}
\value{
El \code{data.frame} codificado
}
\description{
Función tomada del paquete \code{mltools}. Originalmente
    trabaja con \code{data.table} pero en esta versión acepta
    \code{data.frame} y su salida es un \code{data.frame}
}
\examples{
one_hot(iris)
}
