% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_plotly.R
\encoding{UTF-8}
\name{pca_crim}
\alias{pca_crim}
\title{Análisis de componentes principales por años - Delitos}
\usage{
pca_crim(data, years, id_col, date_col, tz_col, val_col)
}
\arguments{
\item{data}{Un \code{data.frame} de delito previamente filtrado}

\item{years}{Los años que deben ser incluidos en el análisis}

\item{id_col}{Carácter. Nombre de la columna que contiene los nombres
de las divisiones espaciales de ínteres}

\item{date_col}{Carácter. Nombre de la columna que contiene las fechas}

\item{tz_col}{Carácter. Nombre de la columna que contiene las zonas
horarias del día. Usualmente es "RANG_DIA"}

\item{val_col}{Carácter. Nombre de la columna que contiene la variable
numérica que se necesita agregar por fechas y divisiones espaciales}
}
\value{
Dos listas: \code{pca_result} contiene los ACP listados por
    año y \code{axis_names} una sugerencia de nombre para los dos
    primeros componentes
}
\description{
Función dirigida a modularización de las Shiny App.
}
\examples{
objeto <- pca_crim(data = test2021::siedco_wc[Delito == "HURTO DE CELULARES",],
years = unique(data.table::year(test2021::siedco_wc[Delito == "HURTO DE CELULARES",FECHA])),
id_col = "NOM_DIV1",
date_col = "FECHA",
tz_col = "RANG_DIA",
val_col = "Numero Hechos")
}
