% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_plotly.R
\encoding{UTF-8}
\name{pca_crime_pre}
\alias{pca_crime_pre}
\title{Procesamiento de datos de delitos para análisis de componentes
    principales}
\usage{
pca_crime_pre(data, id_col, date_col, tz_col, val_col)
}
\arguments{
\item{data}{Un \code{data.frame} de delito previamente filtrado}

\item{id_col}{Carácter. Nombre de la columna que contiene los nombres
de las divisiones espaciales de ínteres}

\item{date_col}{Carácter. Nombre de la columna que contiene las fechas}

\item{tz_col}{Carácter. Nombre de la columna que contiene las zonas
horarias del día. Usualmente es "RANG_DIA"}

\item{val_col}{Carácter. Nombre de la columna que contiene la variable
numérica que se necesita agregar por fechas y divisiones espaciales}
}
\value{
Un \code{data.frame} procesado ideal para ACP
}
\description{
Función dirigida a modularización de las Shiny App.
}
\examples{
# Aumentar los datos de siedco para ACP
pca_crime_pre(data = test2021::siedco_wc[Delito == "HURTO DE CELULARES",],
id_col = "NOM_DIV1",
date_col = "FECHA",
tz_col = "RANG_DIA",
val_col = "Numero Hechos")
}
