% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virids.R
\name{plasma}
\alias{plasma}
\title{Plasma palette}
\usage{
plasma(n, alpha = 1, begin = 0, end = 1, direction = 1)
}
\arguments{
\item{n}{Cantidad de colores}

\item{alpha}{Transparencia}

\item{begin}{Inicio de la paleta}

\item{end}{Fin de la paleta}

\item{direction}{Direccion de la paleta}
}
\description{
Exported object from
}
