% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_mod_selector.R
\encoding{UTF-8}
\name{predict.varest}
\alias{predict.varest}
\title{Método de predicción para objetos de clase varest (principalmente)}
\usage{
\method{predict}{varest}(object, ..., n.ahead = 10, test = NULL, dumvar = NULL)
}
\arguments{
\item{object}{Un objeto de clase \code{varest}}

\item{...}{Argumentos adicionales. No usados}

\item{n.ahead}{Un entero especificando la cantidad de pasos adelante a
predecir. No usado en one-step-ahead forecast}

\item{test}{Matriz de datos con las observaciones de prueba. \code{NULL}
por defecto. De suministrarse se omite \code{n.ahead} y se realiza
el procedimiento one-step-ahead forecast}

\item{dumvar}{Objeto que contiene las variables exógenas (de haber sido
usadas en el ajuste VAR)}
}
\value{
Una lista que contiene
   \describe{
       \item{fcst}{Una matriz cuyas filas representan los puntos futuros
       y las columnas las variables endógenas pronosticadas}
       \item{endog}{Matriz con la muestra - solo variables endógenas}
       \item{model}{El objeto VAR estimado}
       \item{exo.fcst}{Los valores pronosticados de las variables exónegas.
       Si no se suministraron devuelve \code{NULL}. De momento no se puede
       usar para one-step-ahead forecast}
   }
}
\description{
Objeto exportado del paquete \code{vars}. Puede realizar
    pronósticos a objetos de clases \code{vars} y \code{vec2var} (no
    incluidos acá de momento). Su código fuente fue alterado para omitir
    la creación de intervalos de confianza, como a su vez su uso
    para obtener predicciones un paso adelante con información de prueba
    disponible.
}
