% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_shinyMobile.R
\name{preview_mobile}
\alias{preview_mobile}
\title{Exported function from shinyMobile}
\usage{
preview_mobile(
  appPath = NULL,
  url = NULL,
  port = 3838,
  device = c("iphoneX", "galaxyNote8", "iphone8", "iphone8+", "iphone5s", "iphone5c",
    "ipadMini", "iphone4s", "nexus5", "galaxyS5", "htcOne"),
  color = NULL,
  landscape = FALSE
)
}
\arguments{
\item{appPath}{Direccion de la app}

\item{url}{URL de la app}

\item{port}{Puerto}

\item{device}{Tipo de movil a imitar}

\item{color}{Color del movil}

\item{landscape}{Interno}
}
\description{
Exported function from shinyMobile
}
