% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_shinyMobile.R
\encoding{UTF-8}
\name{preview_shinyMobile}
\alias{preview_shinyMobile}
\title{Mapas de crimen interactivos - Previzualisación en dispositivos moviles}
\usage{
preview_shinyMobile(to_run, device)
}
\arguments{
\item{to_run}{Nombre de la carpeta que contienen los archivos de la App.
Llamar la función con cualquier argumento para ver las aplicaciones
disponibles.}

\item{device}{El nombre del dipositivo movil que será emulado}
}
\description{
Las aplicaciones estan descritas en la viñeta
    \code{manual_de_uso}. Esta función es un envolvente a
    \code{preview_mobile} que evita el trabajo de identificar ruta/archivos
    de la aplicación requerida. No obstante es necesario realizar los
    pasos restantes en terminal como se describe
    [acá](https://rinterface.github.io/shinyMobile/articles/shinyMobile_tools.html).
}
