% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_plotly.R
\encoding{UTF-8}
\name{proj_pca}
\alias{proj_pca}
\title{Proyecciones de ACP respecto al primer año - Delitos}
\usage{
proj_pca(x, year_base, n_comp)
}
\arguments{
\item{x}{Una lista con nombres (años) que contiene los resultados
de ACP para cada año}

\item{year_base}{El año tomado como base o inicial. Por defecto
se usa el menor año}

\item{n_comp}{El total de componentes principales a usar en la
proyección}
}
\value{
Una lista con proyecciones de las matrices sobre el primer año
}
\description{
Función dirigida a modularización de las Shiny App.
}
\examples{
objeto <- pca_crim(data = test2021::siedco_wc[Delito == "HURTO DE CELULARES",],
years = unique(data.table::year(test2021::siedco_wc[Delito == "HURTO DE CELULARES",FECHA])),
id_col = "NOM_DIV1",
date_col = "FECHA",
tz_col = "RANG_DIA",
val_col = "Numero Hechos")
objeto_2 <- proj_pca(x = objeto$pca_result, year_base = 2012, n_comp = 4)
}
