% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_models.R
\encoding{UTF-8}
\name{reg_models}
\alias{reg_models}
\title{Modelos de regresión - Mapa de datos}
\usage{
reg_models(
  split_out,
  factor_split,
  dates_col,
  models = c("arima", "arimax", "linear", "arima_linear", "linear_arima", "nn_1hl",
    "nn_2hl", "nn_3hl"),
  only_best = FALSE,
  save_obj = FALSE,
  ...
)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.
Las columnas de factores deben estar en su correspondiente formato.}

\item{factor_split}{Los nombres de las columnas usadas para identificar
los objetos medidos en el tiempo.}

\item{dates_col}{El nombre de la columna que contiene las fechas.}

\item{models}{El listado de modelos a evaluar/retornar}

\item{only_best}{Lógico. Retornar solamente el modelo con el mayor
RMSE en validación}

\item{save_obj}{Lógico. Guardar el objeto retornado en disco duro. Se
seleccionado el directorio de trabajo por defecto}

\item{...}{argumentos adicionales pasados a la función principal o a los modelos}
}
\value{
Una lista que contiene los modelos ajustados y las metrícas por modelo
}
\description{
Prototipo de función principal que integra parte de la pipeline
    de modelos de regresión: selección de modelo, ajuste de modelo y
    reporte de métricas principales. Actualmente la función es capaz de
    procesar los siguientes modelos con sus respectivas especificaciones:
    \describe{
       \item{Modelo lineal:}{Un modelo \code{lm} estándar. Puede incluir
       selección stepwise}
       \item{ARIMA:}{Modelos de series de tiempo lineales ARIMA sin
       componente estacional}
       \item{ARIMAX:}{Modelos ARIMA con variables númericas exogenas}
       \item{ARIMA+Lineal}{Modelos ARIMA complementados con un modelo
       lineal para los residuos derivados}
       \item{Lineal+ARIMA}{Modelos lineales complementados con modelos
       ARIMA para los residuos derivados}
       \item{NN:}{Redes neuronales con \code{keras} de una o dos capas ocultas}
    }
    Las métricas reportadas son las descritas en \code{reg_metrics}

    TODO: crear función reg_metrics y selector NN
}
