% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_map.R
\encoding{UTF-8}
\name{rmse_map}
\alias{rmse_map}
\title{Visualización del RMSE de un modelo por sectores}
\usage{
rmse_map(
  city,
  type = c("arima", "arimax", "linear", "arima_linear", "linear_arima"),
  fit_object,
  merge_on,
  split_out,
  ...
)
}
\arguments{
\item{city}{El objeto donde se guarda la información espacial (geometrías)}

\item{type}{El tipo de modelo usado. Al momento de realización se reconoce
"arima", "arimax", "linear" y "arima_linear"}

\item{fit_object}{El objeto retornado por la salida de ajuste automático
correspondiente}

\item{merge_on}{El nombre de la columna usada para juntar las salidas de
los modelos y el objeto de \code{city}. Debe identificar los agregados
espaciales a visualizar, en el caso de modelos "arima", "arimax" y
"arima_lin" debe ser igual al parámetro \code{factor_split} suministrado.}

\item{split_out}{La partición de los datos en entrenamiento, validación
y prueba}

\item{...}{Otros parámetros adicionales que pueden variar según el tipo
de modelo suministrado. Para todos se requiere un parámetro
\code{names_col} que sea igual al nombre de la columna (dentro del
objeto espacial) que contenga los **labels** deseados de las agregaciones
espaciales.}
}
\value{
Un mapa interactivo que muestra los RMSE
}
\description{
Función experimental usada para representar gráficamente
    en \code{leaflet} los RMSE de validación obtenidos de una salida
    de selección de modelo.

    TODO: Ejemplos.
}
