% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{rolling_cv}
\alias{rolling_cv}
\title{Asistente para validación cruzada para series de tiempo}
\usage{
rolling_cv(Y, X, dates_col, size, horizon, fixed_window = FALSE, ...)
}
\arguments{
\item{Y}{El vector de respuesta. Actualmente no usado, pero será implementado
en el algoritmo para clasificación (integridad de clases)}

\item{X}{La matriz de características. Debe contener una columna con fechas}

\item{dates_col}{El nombre de la columna de fechas en \code{X}}

\item{size}{Entero. La cantidad de periodos usados para entrenamiento}

\item{horizon}{Entero. La cantidad de periodos usados como horizonte de
predicción}

\item{fixed_window}{Lógico. Usar o no una ventana de tamaño \code{size} para
entrenamiento, en caso de no usarse se tomarán todas las observaciones
disponibles}

\item{...}{Argumentos adicionales. Control del delta de tiempo \code{time_delta}}
}
\value{
Una lista de indices \code{train} y \code{val} para usar en algoritmos
    de validación cruzada
}
\description{
Siguiendo a [Hyndman y Athanasopoulos (2013)](https://www.otexts.org/fpp/2/5)
    con esta función se crea proporciona una técnica de partición del conjunto
    de datos en bloques contiguos para la realización de validación cruzada
    en series de tiempo y datos longitudinales/panel.

    NOTA: El tamaño de entrenamiento y del horizonte de predicción pueden
    variar si las fechas están repetidas en el conjunto de datos. Esto se
    realiza para poder trabajar con datos longitudinales/panel.

    Se recomienda en versiones futuras reducir el tamaño de los objetos guardados
    por esta función.
}
\examples{
redux <- siedco_full[Fecha < "2015-01-01",]
cv_op <- rolling_cv(redux$`Numero Hechos`, redux[,-"Numero Hechos"], "Fecha",1510, 300, TRUE)
}
