% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\encoding{UTF-8}
\name{run_shiny}
\alias{run_shiny}
\title{Mapas de crimen interactivos}
\usage{
run_shiny(to_run)
}
\arguments{
\item{to_run}{Nombre de la carpeta que contienen los archivos de la App.
Llamar la función con cualquier argumento para ver las aplicaciones
disponibles.}
}
\description{
Las aplicaciones estan descritas en la viñeta
    \code{manual_de_uso}.
}
