% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{scale_split}
\alias{scale_split}
\title{Escalamiento de variables}
\usage{
scale_split(split_out)
}
\arguments{
\item{split_out}{Lista de partición en entrenamiento, validación y prueba}
}
\value{
Lista de partición con características númericas normalizadas
}
\description{
Ideal para modelos dependientes de el escalado de datos.
    Solamente en las características. No aplica para pruebas
}
\examples{
# Definimos Y y X
Y <- iris["Species"]
X <- subset(iris, select = -Species)
# Realizamos la particion
test_split <- list()
n_train <- 80
test_split$X$X_train <- X[1:80,]
test_split$X$X_val <- X[-(1:80),]
test_split$Y$Y_train <- Y[1:80,]
test_split$Y$Y_val <- Y[-(1:80),]
# Normalizacion
test_split <- scale_split(test_split)
}
