% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_col_crim.R
\docType{data}
\encoding{UTF-8}
\name{siedco_wc}
\alias{siedco_wc}
\title{Reportes completos de delitos de alto impacto en la ciudad de Bogotá,
    desde 2010 - Con coordenadas espaciales}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1670786 rows and 13 columns.
}
\usage{
siedco_wc
}
\description{
Reportes completos de delitos de alto impacto en la ciudad de Bogotá,
    desde 2010 - Con coordenadas espaciales
}
\keyword{datasets}
