% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular_delito_vias.R
\encoding{UTF-8}
\name{simular_delito_vias}
\alias{simular_delito_vias}
\title{Simulador en un sistema de vias un delito.}
\usage{
simular_delito_vias(range_day, sp_obj, ...)
}
\arguments{
\item{range_day}{El inicio y final de la secuencia de días a simular. Deben
tener un formato admisible como "AAAA-MM-DD" o "DD/MM/AAAA".}

\item{sp_obj}{El objeto espacial que contiene la información de vías
(geometrías) con columnas auxiliares de velocidad, rango horario,
día de la semana y nombre de la vía.}

\item{...}{Una serie de parámetros opcionales destinados a configurar
las opciones internas de funciones a un nivel especifico.}
}
\value{
Un \code{data.frame} que contiene las columnas
    \describe{
        \item{LAT}{Coordenada de latitud.}
        \item{LNG}{Coordenada de longitud.}
        \item{HORA}{Rango horario en donde ocurrió el delito. Su nombre
        puede ser alterado.}
        \item{DIA_SEMANA}{El día de la semana donde ocurrió el delito. Su
        nombre puede ser alterado.}
        \item{FECHA}{La fecha exacta donde ocurrió el delito. Sin hora.}
    }
}
\description{
En esta rutina se simula un delito arbitrario en una ciudad
    que cuenta con un sistema vial bien definido, durante una ventana de
    tiempo especificada (diario).

    El algoritmo asigna una cantidad aleatoria uniforme de cada delito en
    un día determinado, luego, se reparte la cantidad obtenida entre
    unos rangos horarios establecidos en función de los pesos asignados
    (deben sumar uno). Dado un día y un rango horario, los delitos son
    asignados a una selección aleatoria de vías la cual viene dada por
    unas categorías de velocidades (e.g. lentas y rápidas) con unas
    probabilidades personalizadas por el usuario (o dadas por defecto).
    Finalmente, dentro de cada vía seleccionada se realiza un muestreo
    uniforme en todo el segmento (no necesariamente una recta, tan solo
    debe unir dos puntos y debe ser continuo).

    La mayor parte de la rutina es automática, pero se puede personalizar
    con el mayor detalle posible.

    *Nota:* La cantidad de delitos por día siguen la misma distribución
    uniforme para todos los días de la semana. Las probabilidades de
    observar un delito en un rango de horas/velocidades son asignadas
    con un criterio subjetivo.
}
\examples{
\donttest{
# Simular un delito generico en las vias principales de Bogota, 2020 a 2021
delito_bog <- simular_delito_vias(c("2020-01-01","2021-01-01"),
    movilidad_marzo_bog, col_factor_rango = "HORA")
}
}
