% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular_variables_adicionales.R
\encoding{UTF-8}
\name{simular_variables_adicionales}
\alias{simular_variables_adicionales}
\title{Adicionar columnas simuladas a un conjunto de datos.}
\usage{
simular_variables_adicionales(df_base, ...)
}
\arguments{
\item{df_base}{El \code{data.frame} al cual se le adicionarán las columnas.}

\item{...}{Múltiples \code{data.frame} de dos columnas, donde la primera
contiene los niveles/categórias de la variable y la segunda, que debe
llevar por nombre PROP, tendrá las probabilidades de cada una (deben
sumar a uno). Es necesario que cada \code{data.frame} lleve un
nombre (key) en el formato \code{df_prop_***}, donde los asteriscos deben
ser reemplazados con el nombre que se espera obtener la nueva columna.}
}
\value{
El objeto \code{df_base} con la adición de las nuevas columnas.
}
\description{
Función auxiliar que realiza un muestreo aleatorio con
    reemplazamiento de una variable categórica cuyos valores y probabilidades
    (PROP) de ocurrencia se encuentran consignadas en un \code{data.frame}.
    La cantidad de muestras tomadas corresponderá al total de filas del
    \code{data.frame} declarado como \code{df_base}.

    TODO: Ejemplos.
}
