% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_crime_cat.R
\encoding{UTF-8}
\name{sp_crime_pre}
\alias{sp_crime_pre}
\title{Pre-procesamiento de para \code{\link{sp_crime_cat}}.}
\usage{
sp_crime_pre(
  city = lon_poly_ward_b,
  geo_div = "DISTRICT",
  var_sum = "HECTARES"
)
}
\arguments{
\item{city}{Un \code{SpatialPolygonsDataFrame} de la ciudad seleccionada.}

\item{geo_div}{Un string que contiene el nombre del campo que determina
la división geo-política de ínteres, el cual debe estar contenido
en el campo \code{data} del \code{city}. Se acepta también vectores
de strings para múltiples llaves a conservar en el resultado.}

\item{var_sum}{Un string que contiene el nombre del campo auxiliar a ser
agregado por la agrupación obtenida por \code{geo_div}. Debe ser
un campo númerico de \code{data} en \code{city}.}
}
\value{
Una lista con un \code{SpatialPolygonsDataFrame} y el nombre
    del campo usado en \code{geo_div}.
}
\description{
Esta función debe ser llamada antes de la ejecución
    del algoritmo principal \code{\link{sp_crime_cat}}. Se deja
    como función independiente para evitar procesos redundantes.
}
\examples{
\dontrun{
sp_crime_pre()
}
}
