% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{split_by_factors}
\alias{split_by_factors}
\title{Separar bases de datos en factores únicos.}
\usage{
split_by_factors(x, factor_split)
}
\arguments{
\item{x}{La base de datos longitudinales a ser particionada. Debe estar en
formato \code{data.frame}.}

\item{factor_split}{Las columnas a usar para crear las particiones.}
}
\value{
Una lista de \code{data.frame} que tienen nivele o combinaciones
    únicas de los factores seleccionados.
}
\description{
Función pensada para datos de tipo longitudinal, la cual se
    destina como uso previo a funciones relacionadas con el tiempo como
    \code{add_laggs}. De momento solo puede trabajar con columnas
    cuyo tipo sea factor o carácter.
}
\examples{
# Crear una secuencia de fechas
X_fecha <- seq(as.Date("1995-03-26"), length = 5, by = "1 day")
# Crear los factores
X_color <- rep(c("Azul", "Verde"), each = 5)
# Crear el data.frame y mezclarlo
X <- data.frame(FECHA = X_fecha, COLOR = X_color)
X <- X[sample(1:10,10),]
# Particion por color
split_by_factors(X,"COLOR")
}
