% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_models.R
\encoding{UTF-8}
\name{split_nn}
\alias{split_nn}
\title{Particiones de datos para redes neuronales - Keras}
\usage{
split_nn(split, ...)
}
\arguments{
\item{split}{Un objeto retornado por \code{create_split_long} o
\code{create_split_ts}}

\item{...}{Argumentos adicionales. \code{expand_y} controla si la
respuesta (categoríca) debe ser expandida en dummies o no, por
defecto se aplica}
}
\description{
Particiones de datos para redes neuronales - Keras
}
