% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_mod_selector.R
\encoding{UTF-8}
\name{svm_mod_selector}
\alias{svm_mod_selector}
\title{Ajuste de maquinas de vectores de soporte para clasificación.}
\usage{
svm_mod_selector(split_out, scaled = FALSE, quiet = FALSE, tuning = FALSE, ...)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.
Las columnas de factores deben estar en su correspondiente formato.}

\item{scaled}{¿Los datos están escalados? De usar falso se realizará
su escalamiento internamente en \code{svm}.}

\item{quiet}{Lógico. Callar las salidas en consola de los modelos.}

\item{tuning}{Lógico. Retornar o no el mejor modelo posible.}

\item{...}{argumentos adicionales pasados a \code{svm} o \code{best.svm}
consultar la función \code{\link[e1071]{svm}} para mayor información.}
}
\value{
Una lista que incluye la mejor SVM (o mejores en multi-clases)
    junto a la **accuracy** reportada en entrenamiento y validación
}
\description{
Se usa la rutina \code{svm} de la librería \code{e1071} para
    crear un clasificador o múltiples, según la naturaleza de la respuesta.
    En caso de múltiples categorías se usa el método uno contra todas con
    selección de predicción mediante votación. Se reporta la **accuracy**
    en entrenamiento y validación.

    Dado que esto es un modelo de clasificación, la respuesta esperada
    debe ser un \code{factor}. Además, por como funciona la rutina
    \code{svm} las columnas numéricas deben estar escaladas
    para asegurar rápida convergencia.

    *Nota* en versiones futuras se incluirá ajuste por CV para el
    pesos de los parámetros (usando \code{caret}, por ejemplo)

    TODO: Ejemplos.
}
