% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_mod_selector.R
\encoding{UTF-8}
\name{VAR}
\alias{VAR}
\title{Estimación de un modelo VAR(p)}
\usage{
VAR(
  y,
  p = 1,
  type = c("const", "trend", "both", "none"),
  season = NULL,
  exogen = NULL,
  lag.max = NULL,
  ic = c("AIC", "HQ", "SC", "FPE"),
  experimental = FALSE
)
}
\arguments{
\item{y}{Datos que contienen la información endógena}

\item{p}{Entero que indica el orden de rezago (1 por defecto)}

\item{type}{Tipo de regresión determinista a incluir}

\item{season}{Inclusión de variables dummy centradas estacionales (entero)}

\item{exogen}{Inclusión de variables exógenas}

\item{lag.max}{Entero, determina el orden máximo de rezago para selección
automática}

\item{ic}{Carácter, selección del criterio de información. \code{lag.max}
debe ser no nulo}

\item{experimental}{Pendiente}
}
\value{
Una lista con atributo de clase \code{varest} que contiene diferentes
    elementos. Consultar documentación de \code{vars} para ampliar
}
\description{
Objeto exportado del paquete \code{vars}.
}
