% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_mod_selector.R
\encoding{UTF-8}
\name{var_mod_selector}
\alias{var_mod_selector}
\title{PENDIENTE}
\usage{
var_mod_selector(
  split_out,
  factor_split,
  dates_col,
  season = NULL,
  xreg_cols = NULL,
  auto_sel = TRUE,
  restriction = NULL,
  quiet = FALSE,
  experimental = FALSE,
  ...
)
}
\arguments{
\item{split_out}{La partición entrenamiento/validación/prueba de los datos.
Las columnas de factores deben estar en su correspondiente formato.}

\item{factor_split}{Carácter. Nombre de la columna que separa las variables
endógenas}

\item{dates_col}{Carácter. Nombre de la columna que contiene las fechas.
No requiere un formato determinado pero requiere que sean las mismas
en todos los niveles de \code{factor_split}}

\item{season}{Entero. La frecuencia estacional (dummy).}

\item{xreg_cols}{Los nombres de las columnas que contienen las
variables exógenas. No usado de momento.}

\item{auto_sel}{Lógico. Realizar selección automática del orden VAR. Por
defecto se busca hasta p = 10. Configurar en \code{...} con las opciones
\code{lag.max} (p máximo) y \code{ic} (criterio de información usado).}

\item{restriction}{Matriz usada para crear restricciones sobre el objecto
estimado. En este caso se recomienda asignar p manualmente}

\item{quiet}{Callar las salidas en consola de los modelos.}

\item{experimental}{Lógico. Restricción usando correlaciones parciales}

\item{...}{Argumentos adicionales}
}
\value{
Una lista que incluye los modelos lineales equivalentes que
    de la estimación del VAR (OLS). La cantidad de modelos de regresión
    es igual al total de variables endógenas declaradas. Se incluye además
    los RMSE totales en entrenamiento y validación.
}
\description{
PENDIENTE

    TODO: Ejemplos
}
