% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular_delito_divisiones.R
\encoding{UTF-8}
\name{var_unidad_divisiones}
\alias{var_unidad_divisiones}
\title{Simulación de varianza 2x2.}
\usage{
var_unidad_divisiones(df_kernels, sd_range = c(0.005, 0.01))
}
\arguments{
\item{df_kernels}{Un \code{data.frame} que contiene la información
correspondiente a los kernels, clasificada por division administrativa.}

\item{sd_range}{El rango entre el cual se obtendran los valores de las
desviaciones estándar.}
}
\value{
Una lista cuyos elementos mínimos son matrices de varianzas 2x2.
    La cantidad de elementos viene determinada por la cantidad de unidades
    y kernels reportadas en \code{df_kernels}.
}
\description{
Función de utilidad usada en \code{simular_delito_divisiones}
    para ser pasada como argumento adicional.

    TODO: Ejemplos.
}
