% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_mod_selector.R
\encoding{UTF-8}
\name{VARselect}
\alias{VARselect}
\title{Criterios de información y FPE para diferentes VAR(p)}
\usage{
VARselect(
  y,
  lag.max = 10,
  type = c("const", "trend", "both", "none"),
  season = NULL,
  exogen = NULL
)
}
\arguments{
\item{y}{Datos que contienen la información endógena}

\item{lag.max}{Entero, determina el orden máximo de rezago para selección
automática}

\item{type}{Tipo de regresión determinista a incluir}

\item{season}{Inclusión de variables dummy centradas estacionales (entero)}

\item{exogen}{Inclusión de variables exógenas}
}
\value{
Una lista. Consultar detalles en \code{vars}
}
\description{
Objeto exportado del paquete \code{vars}.
}
