% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virids.R
\name{viridis}
\alias{viridis}
\title{Viridis main function (inner)}
\usage{
viridis(n, alpha = 1, begin = 0, end = 1, direction = 1, option = "D")
}
\arguments{
\item{n}{Cantidad de colores}

\item{alpha}{Transparencia}

\item{begin}{Inicio de la paleta}

\item{end}{Fin de la paleta}

\item{direction}{Direccion de la paleta}

\item{option}{Interno}
}
\description{
Exported function from viridis package
}
